fix wait time arg

+++ client.py
@@ -197,7 +197,7 @@ async def main():
     parser.add_argument("--frame-interval", type=int, default=25, help="Send every Nth frame (ignored if --frames is used)")
     parser.add_argument("--frames", type=str, default=None, 
                         help="Comma-separated list of specific frame numbers to send (e.g., '100,200,300'). Overrides --frame-interval.")
-    parser.add_argument("--wait-time", type=int, default=1, help="Wait time between frames in seconds")
+    parser.add_argument("--wait-time", type=float, default=1.0, help="Wait time between frames in seconds")
     
     args = parser.parse_args()
